requirements.txt
aiohttp==3.9.3
asyncio==3.4.3
loguru==0.7.2
networkx==3.2.1
numpy==1.26.4
openai==1.12.0
pydantic==2.6.3
pytest==8.0.2
pytest-asyncio==0.23.5
python-dotenv==1.0.1
tenacity==8.2.3
typing-extensions==4.9.0 